% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok_pal.R
\name{ok_pal}
\alias{ok_pal}
\title{ok_pal}
\usage{
ok_pal(pal)
}
\arguments{
\item{pal}{an integer between 1 and 10.}
}
\description{
This function calls a palette based on number.
}
\examples{
ok_pal(1)
}
