% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pal.R
\name{plot_pal}
\alias{plot_pal}
\title{plot_pal}
\usage{
plot_pal(pal)
}
\arguments{
\item{pal}{palette to display; can be the name or an integer between 1 and 10.}
}
\description{
This function plots the colors from a ok palette, and displays the source painting.
}
\examples{
plot_pal(pal_1)
}
